requirements.txt
aiomysql==0.0.22
py-cord==2.0.0b5
python-dotenv==0.19.0